+++ Tests/StreamChatTests/Controllers/ChannelController/ChannelController_Tests.swift
@@ -5792,7 +5792,7 @@ final class ChannelController_Tests: XCTestCase {
         // GIVEN
         let level: PushPreferenceLevel = .all
         let expectedPreference = PushPreference(
-            chatLevel: .all,
+            level: .all,
             disabledUntil: nil
         )
 
@@ -5851,7 +5851,7 @@ final class ChannelController_Tests: XCTestCase {
         // GIVEN
         let date = Date().addingTimeInterval(3600)
         let expectedPreference = PushPreference(
-            chatLevel: .all,
+            level: .all,
             disabledUntil: date
         )