+++ Tests/StreamChatTests/Workers/ChannelUpdater_Tests.swift
@@ -2605,7 +2605,7 @@ final class ChannelUpdater_Tests: XCTestCase {
         let response = UpsertPushPreferencesResponse(
             userChannelPreferences: [
                 "userId": [
-                    cid.rawValue: PushPreference(chatLevel: "all")
+                    cid.rawValue: PushPreference(level: "all")
                 ]
             ],
             userPreferences: [:]
@@ -2697,7 +2697,7 @@ final class ChannelUpdater_Tests: XCTestCase {
         let response = UpsertPushPreferencesResponse(
             userChannelPreferences: [
                 "userId": [
-                    cid.rawValue: PushPreference(chatLevel: "mentions")
+                    cid.rawValue: PushPreference(level: "mentions")
                 ]
             ],
             userPreferences: [:]
@@ -2740,7 +2740,7 @@ final class ChannelUpdater_Tests: XCTestCase {
             let pushPreferenceDTO = try session.savePushPreference(
                 id: cid.rawValue,
                 payload: .init(
-                    chatLevel: "none",
+                    level: "none",
                     disabledUntil: nil
                 )
             )
@@ -2759,7 +2759,7 @@ final class ChannelUpdater_Tests: XCTestCase {
         let response = UpsertPushPreferencesResponse(
             userChannelPreferences: [
                 "userId": [
-                    cid.rawValue: PushPreference(chatLevel: "all")
+                    cid.rawValue: PushPreference(level: "all")
                 ]
             ],
             userPreferences: [:]